refactor: moved autocomplete import
