+++ src/components/AutocompleteInput/AutocompleteInput.stories.tsx
@@ -1,7 +1,7 @@
 // also exported from '@storybook/react' if you can deal with breaking changes in 6.1
 import { Meta, Story } from '@storybook/react/types-6-0';
 import React from 'react';
-import { AutocompleteInput, AutocompleteInputProps } from '.';
+import { AutocompleteInput, AutocompleteInputProps } from './AutocompleteInput';
 import { MenuIcon } from '../Icons';
 
 export default {
